Lab 12 Task/Academic Bot/requirements.txt
openai 
sentence-transformers
faiss-cpu 
PyMuPDF 
flask 
flask-cors
